Add a date range slider widget
If WPS input parameter is of type string and the identifier starts with
"date-slider" then a date range slider widget is used. If the input
parameter default value is of the form "yyyy/mm/dd|yyyy/mm/dd" then the
two numbers are used as the beginning and end of the range.