Updated mongodb lookup on "title" as well as "identifier"
Modified `MongodbCatalog.get_record_by_id(self, identifier)`

It now gets a database record by comparing the input parameter with
the "identifier" or "title".

If the identifier does not match, then a case-insensitive
match is made on the title (replacing "_" with " "). This
enables _deterministic_ URLs in which the WPS can be identified
by a meaningful string as well as the standard long identifier.