File Upload Widget
This makes use of the deform `FileUploadWidget`.

The implementation of the `FileUploadTempStore` writes the file to disk.
The root used for the path to write the files is taken from the
configuration variable `file-upload-storage-dir`. The limit of the file
size that can be uploaded is set by the configuration variable
`file-upload-max-size-mb`

The implementation of the `FileUploadWidget` returns a string that is
the path to the uploaded file.