Changes have been made for where user management is handled outside of Phoenix.
This relates to #23

A new variable has been introduced, `local-user-management`, which can
be set in the `custom.cfg`, e.g. `local-user-management = false`. By
default the value is `true` and the current behaviour remains unchanged.

When set to `false` then a number of end-points will not be found:
- people
- people/profile/{userid}/{tab}
- people/delete/{userid}
These items will also be removed from the navbar